requirements.txt
lxml >= 4.6.3
Sickle >= 0.7.0
flake8 >= 3.9.2
black >= 21.7b0
pytest >= 6.2.4
intake >= 0.6.2
aiohttp >= 3.7.4.post0
jsonpath_ng >= 1.5.3
apache-airflow
